requirements-test.txt
pytest==8.1.1
pytest-mock==3.14.0
pytest-cov==5.0.0
moto[dynamodb]==4.2.14
responses==0.25.0
freezegun==1.4.0
boto3==1.34.84
botocore==1.34.84
PyJWT==2.8.0
cryptography==42.0.5
